+++ VkConnector.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26430.16
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Telegram_VkMessenger.Api", "Telegram_VkMessenger.Api\Telegram_VkMessenger.Api.csproj", "{82A93E77-F859-4E19-A09F-29FAE1251AD2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VkConnector", "VkConnector\VkConnector.csproj", "{31192FDE-6A64-4BB9-88C7-E10CB77C6B3E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{82A93E77-F859-4E19-A09F-29FAE1251AD2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{82A93E77-F859-4E19-A09F-29FAE1251AD2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{82A93E77-F859-4E19-A09F-29FAE1251AD2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{82A93E77-F859-4E19-A09F-29FAE1251AD2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{31192FDE-6A64-4BB9-88C7-E10CB77C6B3E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{31192FDE-6A64-4BB9-88C7-E10CB77C6B3E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{31192FDE-6A64-4BB9-88C7-E10CB77C6B3E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{31192FDE-6A64-4BB9-88C7-E10CB77C6B3E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
