+++ VkConnector/Properties/launchSettings.json
