+++ VkConnector/VkConnector.csproj
