+++ VkConnector/appsettings.Development.json
