+++ VkConnector/appsettings.json
